+++ app/src/main/cpp/module.cpp
@@ -1,14 +1,18 @@
 #include <android/log.h>
 #include <sys/system_properties.h>
 #include <unistd.h>
+#include <filesystem>
 
 #include "zygisk.hpp"
 #include "dobby.h"
-#include "json.hpp"
 
 #define LOGD(...) __android_log_print(ANDROID_LOG_DEBUG, "PIF/Native", __VA_ARGS__)
 
-static std::string SECURITY_PATCH, FIRST_API_LEVEL, VNDK_VERSION, BUILD_ID;
+#define CLASSES_DEX "/data/adb/modules/playintegrityfix/classes.dex"
+
+#define PIF_JSON "/data/adb/pif.json"
+
+#define PIF_PROP "/data/adb/pif.prop"
 
 typedef void (*T_Callback)(void *, const char *, const char *, uint32_t);
 
@@ -16,29 +20,13 @@ static volatile T_Callback o_callback = nullptr;
 
 static void modify_callback(void *cookie, const char *name, const char *value, uint32_t serial) {
 
-    if (cookie == nullptr || name == nullptr || value == nullptr || o_callback == nullptr) return;
+    if (cookie == nullptr || name == nullptr || o_callback == nullptr) return;
 
-    std::string_view prop(name);
+    if (std::string_view(name).ends_with("api_level")) {
 
-    if (prop.ends_with("security_patch")) {
-        if (!SECURITY_PATCH.empty()) {
-            value = SECURITY_PATCH.c_str();
-        }
-    } else if (prop.ends_with("api_level")) {
-        if (!FIRST_API_LEVEL.empty()) {
-            value = FIRST_API_LEVEL.c_str();
-        }
-    } else if (prop.ends_with("vndk.version")) {
-        if (!VNDK_VERSION.empty()) {
-            value = VNDK_VERSION.c_str();
-        }
-    } else if (prop == "ro.build.id") {
-        if (!BUILD_ID.empty()) {
-            value = BUILD_ID.c_str();
-        }
+        value = "23";
+        LOGD("Set '%s' to '%s'", name, value);
     }
-
-    if (!prop.starts_with("cache") && !prop.starts_with("debug")) LOGD("[%s]: %s", name, value);
 
     return o_callback(cookie, name, value, serial);
 }
@@ -85,103 +73,53 @@ public:
 
             if (strcmp(name, "com.google.android.gms.unstable") == 0) {
 
+                auto rawDir = env->GetStringUTFChars(args->app_data_dir, nullptr);
+                dir = rawDir;
+                env->ReleaseStringUTFChars(args->app_data_dir, rawDir);
+
+                long size = dir.size();
+                bool done = false;
+
                 int fd = api->connectCompanion();
 
-                read(fd, &dexSize, sizeof(dexSize));
-                read(fd, &jsonSize, sizeof(jsonSize));
+                write(fd, &size, sizeof(long));
+                write(fd, dir.data(), size);
 
-                if (dexSize < 1 || jsonSize < 1) {
-
-                    LOGD("Couldn't read files in memory!");
-                    api->setOption(zygisk::DLCLOSE_MODULE_LIBRARY);
-
-                } else {
-
-                    auto vectorSize = dexSize + jsonSize;
-
-                    if (vectorSize > 0) {
-                        vector.resize(vectorSize);
-                        read(fd, vector.data(), vectorSize);
-                    }
-                }
+                read(fd, &done, sizeof(bool));
 
                 close(fd);
 
-            } else api->setOption(zygisk::DLCLOSE_MODULE_LIBRARY);
+                LOGD("Files copied: %d", done);
 
-        } else api->setOption(zygisk::DLCLOSE_MODULE_LIBRARY);
+                goto clear;
+            }
+        }
 
+        api->setOption(zygisk::DLCLOSE_MODULE_LIBRARY);
+
+        clear:
         env->ReleaseStringUTFChars(args->nice_name, name);
     }
 
     void postAppSpecialize(const zygisk::AppSpecializeArgs *args) override {
-        if (vector.empty()) return;
-
-        LOGD("Read %ld bytes of classes.dex!", dexSize);
-        LOGD("Read %ld bytes of pif.json!", jsonSize);
-
-        if (jsonSize > 0) parseJson();
+        if (dir.empty()) return;
 
         doHook();
 
-        injectDex();
+        std::string classesDex(dir + "/classes.dex");
 
-        vector.clear();
-        json.clear();
-    }
+        FILE *dexFile = fopen(classesDex.c_str(), "rb");
 
-    void preServerSpecialize(zygisk::ServerSpecializeArgs *args) override {
-        api->setOption(zygisk::DLCLOSE_MODULE_LIBRARY);
-    }
+        if (dexFile == nullptr) {
 
-private:
-    zygisk::Api *api = nullptr;
-    JNIEnv *env = nullptr;
-    long dexSize = 0, jsonSize = 0;
-    std::vector<char> vector;
-    nlohmann::json json;
-
-    void parseJson() {
-        std::string_view data(vector.cbegin() + dexSize, vector.cend());
-        json = nlohmann::json::parse(data, nullptr, false, true);
-
-        if (json.contains("SECURITY_PATCH")) {
-            if (json["SECURITY_PATCH"].is_null() || json["SECURITY_PATCH"].empty()) {
-                LOGD("SECURITY_PATCH is null or empty");
-            } else {
-                SECURITY_PATCH = json["SECURITY_PATCH"].get<std::string>();
-            }
+            LOGD("classes.dex doesn't exist... This is weird.");
+            dir.clear();
+            api->setOption(zygisk::DLCLOSE_MODULE_LIBRARY);
+            return;
         }
 
-        if (json.contains("FIRST_API_LEVEL")) {
-            if (json["FIRST_API_LEVEL"].is_null() || json["FIRST_API_LEVEL"].empty()) {
-                LOGD("FIRST_API_LEVEL is null or empty");
-            } else {
-                FIRST_API_LEVEL = json["FIRST_API_LEVEL"].get<std::string>();
-            }
-            json.erase("FIRST_API_LEVEL");
-        }
+        fclose(dexFile);
 
-        if (json.contains("BUILD_ID")) {
-            if (json["BUILD_ID"].is_null() || json["BUILD_ID"].empty()) {
-                LOGD("BUILD_ID is null or empty");
-            } else {
-                BUILD_ID = json["BUILD_ID"].get<std::string>();
-            }
-            json.erase("BUILD_ID");
-        }
-
-        if (json.contains("VNDK_VERSION")) {
-            if (json["VNDK_VERSION"].is_null() || json["VNDK_VERSION"].empty()) {
-                LOGD("VNDK_VERSION is null or empty");
-            } else {
-                VNDK_VERSION = json["VNDK_VERSION"].get<std::string>();
-            }
-            json.erase("VNDK_VERSION");
-        }
-    }
-
-    void injectDex() {
         LOGD("get system classloader");
         auto clClass = env->FindClass("java/lang/ClassLoader");
         auto getSystemClassLoader = env->GetStaticMethodID(clClass, "getSystemClassLoader",
@@ -189,11 +127,11 @@ private:
         auto systemClassLoader = env->CallStaticObjectMethod(clClass, getSystemClassLoader);
 
         LOGD("create class loader");
-        auto dexClClass = env->FindClass("dalvik/system/InMemoryDexClassLoader");
+        auto dexClClass = env->FindClass("dalvik/system/PathClassLoader");
         auto dexClInit = env->GetMethodID(dexClClass, "<init>",
-                                          "(Ljava/nio/ByteBuffer;Ljava/lang/ClassLoader;)V");
-        auto buff = env->NewDirectByteBuffer(vector.data(), dexSize);
-        auto dexCl = env->NewObject(dexClClass, dexClInit, buff, systemClassLoader);
+                                          "(Ljava/lang/String;Ljava/lang/ClassLoader;)V");
+        auto dexStr = env->NewStringUTF(classesDex.c_str());
+        auto dexCl = env->NewObject(dexClClass, dexClInit, dexStr, systemClassLoader);
 
         LOGD("load class");
         auto loadClass = env->GetMethodID(clClass, "loadClass",
@@ -204,48 +142,93 @@ private:
         auto entryClass = (jclass) entryClassObj;
 
         LOGD("call init");
-        auto str = env->NewStringUTF(json.dump().c_str());
         auto entryInit = env->GetStaticMethodID(entryClass, "init", "(Ljava/lang/String;)V");
+        auto str = env->NewStringUTF(findFileInDirectory().c_str());
         env->CallStaticVoidMethod(entryClass, entryInit, str);
+
+        dir.clear();
+    }
+
+    void preServerSpecialize(zygisk::ServerSpecializeArgs *args) override {
+        api->setOption(zygisk::DLCLOSE_MODULE_LIBRARY);
+    }
+
+private:
+    zygisk::Api *api = nullptr;
+    JNIEnv *env = nullptr;
+    std::string dir;
+
+    std::string findFileInDirectory() {
+
+        std::string jsonFilePath = dir + "/pif.json";
+        std::string propFilePath = dir + "/pif.prop";
+
+        FILE *jsonFile = fopen(jsonFilePath.c_str(), "r");
+        if (jsonFile) {
+            fclose(jsonFile);
+            return jsonFilePath;
+        }
+
+        FILE *propFile = fopen(propFilePath.c_str(), "r");
+        if (propFile) {
+            fclose(propFile);
+            return propFilePath;
+        }
+
+        return dir;
     }
 };
 
 static void companion(int fd) {
-    long dexSize = 0, jsonSize = 0;
-    std::vector<char> vector;
 
-    FILE *dex = fopen("/data/adb/modules/playintegrityfix/classes.dex", "rb");
+    long size = 0;
+    std::string dir;
 
-    if (dex) {
-        fseek(dex, 0, SEEK_END);
-        dexSize = ftell(dex);
-        fseek(dex, 0, SEEK_SET);
+    read(fd, &size, sizeof(long));
 
-        vector.resize(dexSize);
-        fread(vector.data(), 1, dexSize, dex);
+    dir.resize(size);
 
-        fclose(dex);
+    read(fd, dir.data(), size);
+
+    LOGD("[ROOT] GMS dir: %s", dir.c_str());
+
+    std::string classesDex(dir + "/classes.dex");
+
+    bool a = std::filesystem::copy_file(CLASSES_DEX, classesDex,
+                                        std::filesystem::copy_options::overwrite_existing);
+
+    std::filesystem::permissions(classesDex, std::filesystem::perms::owner_read |
+                                             std::filesystem::perms::group_read |
+                                             std::filesystem::perms::others_read);
+
+    bool b = false;
+
+    if (std::filesystem::exists(PIF_JSON)) {
+
+        std::string pifJson(dir + "/pif.json");
+
+        b = std::filesystem::copy_file(PIF_JSON, pifJson,
+                                       std::filesystem::copy_options::overwrite_existing);
+
+        std::filesystem::permissions(pifJson, std::filesystem::perms::owner_read |
+                                              std::filesystem::perms::group_read |
+                                              std::filesystem::perms::others_read);
+
+    } else if (std::filesystem::exists(PIF_PROP)) {
+
+        std::string pifProp(dir + "/pif.prop");
+
+        b = std::filesystem::copy_file(PIF_PROP, pifProp,
+                                       std::filesystem::copy_options::overwrite_existing);
+
+        std::filesystem::permissions(pifProp, std::filesystem::perms::owner_read |
+                                              std::filesystem::perms::group_read |
+                                              std::filesystem::perms::others_read);
     }
 
-    FILE *json = fopen("/data/adb/pif.json", "rb");
+    bool done = a && b;
 
-    if (json == nullptr) json = fopen("/data/adb/modules/playintegrityfix/pif.json", "rb");
-
-    if (json) {
-        fseek(json, 0, SEEK_END);
-        jsonSize = ftell(json);
-        fseek(json, 0, SEEK_SET);
-
-        vector.resize(dexSize + jsonSize);
-        fread(vector.data() + dexSize, 1, jsonSize, json);
-
-        fclose(json);
-    }
-
-    write(fd, &dexSize, sizeof(dexSize));
-    write(fd, &jsonSize, sizeof(jsonSize));
-
-    write(fd, vector.data(), vector.size());
+    write(fd, &done, sizeof(bool));
 }
 
 REGISTER_ZYGISK_MODULE(PlayIntegrityFix)