+++ app/src/main/java/es/chiteroman/playintegrityfix/EntryPoint.java
@@ -4,7 +4,7 @@ import android.os.Build;
 import android.util.JsonReader;
 import android.util.Log;
 
-import java.io.StringReader;
+import java.io.FileReader;
 import java.lang.reflect.Field;
 import java.security.KeyStore;
 import java.security.KeyStoreSpi;
@@ -14,21 +14,39 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
+import java.util.Properties;
 
 public class EntryPoint {
     private static final Map<String, String> map = new HashMap<>();
 
-    public static void init(String data) {
-        try (JsonReader reader = new JsonReader(new StringReader(data))) {
-            reader.beginObject();
-            while (reader.hasNext()) {
-                String key = reader.nextName();
-                String value = reader.nextString();
-                map.put(key, value);
+    public static void init(String file) {
+
+        if (file.endsWith(".json")) {
+
+            try (JsonReader reader = new JsonReader(new FileReader(file))) {
+                reader.beginObject();
+                while (reader.hasNext()) {
+                    String key = reader.nextName();
+                    String value = reader.nextString();
+                    map.put(key, value);
+                }
+                reader.endObject();
+            } catch (Exception e) {
+                LOG("Error parsing JSON file: " + e);
             }
-            reader.endObject();
-        } catch (Exception e) {
-            LOG("Couldn't parse JSON from Zygisk lib: " + e);
+
+        } else if (file.endsWith(".prop")) {
+
+            try {
+                Properties properties = new Properties();
+
+                properties.load(new FileReader(file));
+
+                properties.forEach((o, o2) -> map.put((String) o, (String) o2));
+
+            } catch (Exception e) {
+                LOG("Error parsing PROP file: " + e);
+            }
         }
 
         LOG("Map info (keys and values):");
@@ -47,7 +65,7 @@ public class EntryPoint {
     }
 
     protected static boolean isDroidGuard() {
-        return Arrays.stream(Thread.currentThread().getStackTrace()).anyMatch(e -> e.getClassName().toLowerCase(Locale.ENGLISH).contains("droidguard"));
+        return Arrays.stream(Thread.currentThread().getStackTrace()).anyMatch(e -> e.getClassName().toLowerCase(Locale.US).contains("droidguard"));
     }
 
     private static void spoofProvider() {