+++ changelog.md
@@ -2,6 +2,10 @@ We have a Telegram channel!
 If you want to share your knowledge join:
 https://t.me/playintegrityfix
 
-# v14.6
+# v14.7
 
-Updated fingerprint
+- Removed JSON from C++ code.
+- Always spoof api_level props to 23.
+- Added support for reading .prop files like older modules.
+- Removed useless code.
+- Removed default fingerprint, you must include valid props into json or prop file in /data/adb to pass DEVICE verdict.